+++ tests/test_visitors/conftest.py
@@ -19,7 +19,7 @@ _ERROR_FORMAT: Final = ': {0}'
 
 def _produce_error_message(visitor: BaseVisitor) -> str:  # pragma: no cover
     if isinstance(visitor, BaseNodeVisitor):
-        return node_to_string(visitor.tree)
+        return f'\n{node_to_string(visitor.tree)}'
     return ''